2010-05-27	Ralf Corsépius <ralf.corsepius@rtems.org>
	* libfs/src/devfs/devfs.h, libfs/src/devfs/devfs_eval.c:
	Use size_t instead of int for pathlen.

+++ cpukit/libfs/src/devfs/devfs_eval.c
@@ -17,7 +17,7 @@
 
 int devFS_evaluate_path(
   const char                        *pathname,
-  int                                pathnamelen,
+  size_t                             pathnamelen,
   int                                flags,
   rtems_filesystem_location_info_t  *pathloc
 )
